refactor: Split module 'main' into separate modules
Closes #8
