+++ src/pyratings/__init__.py
@@ -1,31 +1,46 @@
-"""
-Copyright 2022 HSBC Global Asset Management (Deutschland) GmbH
+# Copyright 2022 HSBC Global Asset Management (Deutschland) GmbH
+#
+#    Licensed under the Apache License, Version 2.0 (the "License");
+#    you may not use this file except in compliance with the License.
+#    You may obtain a copy of the License at
+#
+#        https://www.apache.org/licenses/LICENSE-2.0
+#
+#    Unless required by applicable law or agreed to in writing, software
+#    distributed under the License is distributed on an "AS IS" BASIS,
+#    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+#    See the License for the specific language governing permissions and
+#    limitations under the License.
 
-   Licensed under the Apache License, Version 2.0 (the "License");
-   you may not use this file except in compliance with the License.
-   You may obtain a copy of the License at
-
-       http://www.apache.org/licenses/LICENSE-2.0
-
-   Unless required by applicable law or agreed to in writing, software
-   distributed under the License is distributed on an "AS IS" BASIS,
-   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-   See the License for the specific language governing permissions and
-   limitations under the License.
-"""
-from pyratings.main import (
-    get_pure_ratings,
+from pyratings.aggregate import get_weighted_average
+from pyratings.clean import get_pure_ratings
+from pyratings.consolidate import (
     get_best_ratings,
     get_second_best_ratings,
     get_worst_ratings,
-    get_scores_from_ratings,
-    get_scores_from_warf,
-    get_ratings_from_scores,
-    get_ratings_from_warf,
-    get_warf_from_scores,
-    get_warf_from_ratings,
-    get_weighted_average,
-    get_warf_buffer,
+)
+from pyratings.get_ratings import get_ratings_from_scores, get_ratings_from_warf
+from pyratings.get_scores import get_scores_from_ratings, get_scores_from_warf
+from pyratings.get_warf import get_warf_from_ratings, get_warf_from_scores
+from pyratings.utils import (
     _assert_rating_provider,
     _extract_rating_provider,
+    _get_translation_dict,
 )
+from pyratings.warf import get_warf_buffer
+
+# define public functions
+__all__ = [
+    "get_best_ratings",
+    "get_pure_ratings",
+    "get_ratings_from_scores",
+    "get_ratings_from_warf",
+    "get_scores_from_ratings",
+    "get_scores_from_warf",
+    "get_second_best_ratings",
+    "get_warf_buffer",
+    "get_warf_from_ratings",
+    "get_warf_from_scores",
+    "get_weighted_average",
+    "get_worst_ratings",
+]
